Refactor: 파일 위치 변경 및 테스트용 코드 제거
