+++ app/src/main/AndroidManifest.xml
@@ -41,7 +41,7 @@
         </intent-filter>-->
         </activity>
         <service
-            android:name=".firebase.cloudMessaging.KFfirebaseMessagingService"
+            android:name=".service.KFfirebaseMessagingService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
